backend/requirements.txt
django
httpx
pytest
psycopg2-binary
djangorestframework
django-cors-headers
djangorestframework-simplejwt
dj-database-url